Move All Worker Code to The Worker Module
